refactor: verify routes
