+++ admin/index.js
@@ -1,36 +1,17 @@
-const rateLimit = require("express-rate-limit");
 const express = require("express");
-const fb = require("./config");
-const auth = fb.auth();
 const app = express();
 const cors = require("cors");
-
+const verifyRouter = require("./routes/verify");
+const mailRouter = require("./routes/mail");
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(cors());
-const limiter = rateLimit({
-  max: 300,
-  windowMs: 60 * 60 * 1000,
-  message: "Too many request from this IP",
-});
-app.use(limiter);
+app.set("view engine", "ejs");
 app.get("/", (req, res) => {
   res.send("Welcome to T-Shirt Design API 🪅");
 });
-app.post("/v1/verify", async (req, res) => {
-  const { access_token, uid } = req.body;
-  console.log(req.body);
-  const response = await auth
-    .verifyIdToken(access_token)
-    .catch((err) => console.log(err));
-  console.log(response);
-  if (response?.uid == uid) {
-    res.send({ isVerified: true });
-  } else {
-    res.send({ isVerified: false });
-  }
-});
-
+app.use("/v1/verify", verifyRouter);
+app.use("/v1/mail", mailRouter);
 app.listen(3000, () => {
   console.log("Running..");
 });
